mod: AddVaccine page and AddVaccineViewModel moved to its folder
