+++ VaxineApp/VaxineApp/Views/Home/Status/Vaccine/AddVaccinePage.xaml
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="VaxineApp.Views.Home.Status.AddVaccinePage"
+             x:Class="VaxineApp.Views.Home.Status.Vaccine.AddVaccinePage"
              Shell.TabBarIsVisible="False"
              Shell.PresentationMode="Modal"
              Title="Add Vaccine">
