+++ VaxineApp/VaxineApp/Views/Home/Status/Vaccine/AddVaccinePage.xaml.cs
@@ -6,10 +6,11 @@ using System.Text;
 using System.Threading.Tasks;
 using VaxineApp.Models;
 using VaxineApp.ViewModels.Home.Status;
+using VaxineApp.ViewModels.Home.Status.Vaccine;
 using Xamarin.Forms;
 using Xamarin.Forms.Xaml;
 
-namespace VaxineApp.Views.Home.Status
+namespace VaxineApp.Views.Home.Status.Vaccine
 {
     [QueryProperty(nameof(Child), nameof(Child))]
     [XamlCompilation(XamlCompilationOptions.Compile)]
